takeoff_agent/requirements.txt
fastapi
uvicorn[standard]
opencv-python-headless
numpy
Pillow
pdf2image
shapely
pyyaml
paddleocr
torch
ultralytics
